Corrige local do troco em pagamentos conforme Schema XML
